implemented get_producers function #369

+++ plugins/chain_api_plugin/include/eosio/chain_api_plugin/chain_api_plugin_results.hpp
@@ -125,7 +125,7 @@ namespace eosio {
 
     struct get_producers_result {
         std::vector<fc::variant> rows; ///< one row per item, either encoded as hex string or JSON object
-        double                   total_producer_vote_weight;
+        uint64_t                  total_producer_vote_weight;
         std::string              more; ///< fill lower_bound with this value to fetch more rows
     };
